adjust folder structure to enforce clean architecture
